Fix native Linux build with -Werror on modern GCC
linux.mk compiles with -Wall -Wextra -Werror. GCC 15 is stricter than the
mingw toolchain used by CI, so the native Linux build fails on warnings the
Windows build never surfaces. All are fixed at the source; no warnings are
suppressed.

unused-but-set-variable:
- Remove dead per-line/iteration counters that were incremented but never
  read (ariadne.c, config_creature.c, config_keeperfx.c, lua_api.c,
  custom_sprites.c pixel counters, player_compevents.c, main.cpp
  total_play_turns).
- Variables used only by debug logging (config_campaigns file counts,
  main.cpp playtime/endtime) are kept but guarded with
  #if (BFDEBUG_LEVEL > 0), matching the existing convention, so they stay
  available in debug builds.

discarded-qualifiers:
- strchr/strrchr yield a pointer into a const string; keep it in a const
  pointer (bflib_fileio.c, centitoml/toml_conv.c) or take the mutable copy
  from the non-const buffer instead (custom_sprites.c).

format:
- Replace the MSVC-only %I64d modifier (parsed as the glibc %I flag on
  Linux, producing wrong output / a format error) with portable %lld and a
  (long long) cast (dungeon_data.c, lvl_script_commands.c).

(cherry picked from commit 3dab5f111edf9521474a3351a7e359416d1ce87b)

+++ src/lvl_script_commands.c
@@ -5127,15 +5127,15 @@ static void set_power_configuration_check(const struct ScriptLine *scline)
     {
         if ( (powervar == 5) && (value->chars[3] != -1) )
         {
-            SCRIPTDBG(7, "Toggling %s castability flag: %I64d", powername, number_value);
+            SCRIPTDBG(7, "Toggling %s castability flag: %lld", powername, (long long)number_value);
         }
         else if ( (powervar == 14) && (value->chars[3] != -1) )
         {
-            SCRIPTDBG(7, "Toggling %s property flag: %I64d", powername, number_value);
+            SCRIPTDBG(7, "Toggling %s property flag: %lld", powername, (long long)number_value);
         }
         else
         {
-            SCRIPTDBG(7, "Setting power %s property %s to %I64d", powername, property, number_value);
+            SCRIPTDBG(7, "Setting power %s property %s to %lld", powername, property, (long long)number_value);
         }
     }
     #endif
